collector/requirements.txt
gqlalchemy==1.4.0
pymongo==4.6.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9